node_subtensor::client

Type Alias HostFunctions

pub type HostFunctions = (SubstrateHostFunctions, HostFunctions, HostFunctions);

Always enable runtime benchmark host functions, the genesis state was built with them so we’re stuck with them forever.

They’re just a noop, never actually get used if the runtime was not compiled with runtime-benchmarks.