node_subtensor

Module rpc

A collection of node-specific RPC methods. Substrate provides the sc-rpc crate, which defines the core RPC layer used by Substrate nodes. This file extends those RPC definitions with capabilities that are specific to this project’s runtime configuration.

Structs

DefaultEthConfig

Default Eth RPC configuration

EthBlockDataCacheTask

Manage LRU caches for block data and their transaction statuses. These are large and take a lot of time to fetch from the database. Storing them in an LRU cache will allow to reduce database accesses when many subsequent requests are related to the same blocks.

EthDeps

Extra dependencies for Ethereum compatibility.

FullDeps

Full client dependencies.

Functions

create_full

Instantiate all full RPC extensions.

Type Aliases

FeeHistoryCache

FeeHistoryCacheLimit

Maximum fee history cache size.

FilterPool

On-memory stored filters created through the eth_newFilter RPC.